requirements.txt
python-docx>=0.8.11
python-dotenv>=0.19.0
openai>=1.0.0
docx2pdf>=0.1.8
argparse>=1.4.0 
pypandoc>=1.11 